# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradeBox HYIP Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List available shipments
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, electronics, agricultural, minerals, textiles, pharma]
        - name: riskGrade
          in: query
          schema:
            type: string
            enum: [A, B, C, D]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, funded, in_transit, delivered]
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get shipment detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
        "404":
          description: Not found
...
  /admin/deposits/{id}/approve:
    post:
      operationId: adminApproveDeposit
      tags: [admin]
      summary: Admin — approve deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDeposit"
  /admin/deposits/{id}/reject:
    post:
      operationId: adminRejectDeposit
      tags: [admin]
      summary: Admin — reject deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDeposit"
...
  /admin/withdrawals/{id}/process:
    post:
      operationId: adminProcessWithdrawal
      tags: [admin]
      summary: Admin — mark withdrawal as processed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProcessWithdrawalInput"
      responses:
        "200":
          description: Processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWithdrawal"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin — list users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: kycStatus
          in: query
          schema:
            type: string
            enum: [all, pending, approved, rejected]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Admin — get user detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/credit-profit:
    post:
      operationId: adminCreditProfit
      tags: [admin]
      summary: Admin — manually credit profit to user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreditProfitInput"
      responses:
        "200":
          description: Profit credited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
  /admin/kyc/{id}/approve:
    post:
      operationId: adminApproveKyc
      tags: [admin]
      summary: Admin — approve KYC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminKyc"
  /admin/kyc/{id}/reject:
    post:
      operationId: adminRejectKyc
      tags: [admin]
      summary: Admin — reject KYC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminKyc"
  /admin/shipments:
    get:
      operationId: adminListShipments
      tags: [admin]
      summary: Admin — list all shipments
      responses:
        "200":
          description: All shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShipmentDetail"

    post:
      operationId: adminCreateShipment
      tags: [admin]
      summary: Admin — create shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentInput"
      responses:
        "201":
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
  /admin/shipments/{id}:
    patch:
      operationId: adminUpdateShipment
      tags: [admin]
      summary: Admin — update shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
...
  /admin/crypto-wallets:
    get:
      operationId: adminGetCryptoWallets
      tags: [admin]
      summary: Admin — get platform deposit wallet addresses
      responses:
        "200":
          description: Wallet addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CryptoAddress"

    patch:
      operationId: adminUpdateCryptoWallets
      tags: [admin]
      summary: Admin — update platform deposit wallet addresses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CryptoWalletsUpdate"
      responses:
        "200":
          description: Updated wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CryptoAddress"
...
components:
  schemas:
...
    Shipment:
      type: object
      required: [id, title, cargoType, origin, destination, profitPercent, riskGrade, fundingGoal, fundingRaised, minInvestment, departureDate, arrivalDate, transitDays, status, freightForwarder, vesselName]
      properties:
        id:
          type: integer
        title:
          type: string
        cargoType:
          type: string
          enum: [electronics, cocoa, lithium, coffee, textiles, pharmaceuticals, cotton, steel]
        origin:
          type: string
        destination:
          type: string
        originCoords:
          type: ["string", "null"]
        destinationCoords:
          type: ["string", "null"]
        profitPercent:
          type: number
        riskGrade:
          type: string
          enum: [A, B, C, D]
        fundingGoal:
          type: number
        fundingRaised:
          type: number
        minInvestment:
          type: number
        departureDate:
          type: string
        arrivalDate:
          type: string
        transitDays:
          type: integer
        status:
          type: string
          enum: [open, funded, in_transit, delivered]
        freightForwarder:
          type: string
        vesselName:
          type: string
        createdAt:
          type: string
    ShipmentDetail:
      allOf:
        - $ref: "#/components/schemas/Shipment"
        - type: object
          properties:
            hsCode:
              type: ["string", "null"]
            weightTons:
              type: ["number", "null"]
            volumeCbm:
              type: ["number", "null"]
            description:
              type: ["string", "null"]
            investorCount:
              type: integer
            myInvestment:
              type: ["number", "null"]
...
    AdminDeposit:
      type: object
      required: [id, userId, traderId, email, coin, amount, txid, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        traderId:
          type: string
        email:
          type: string
        coin:
          type: string
        amount:
          type: number
        txid:
          type: string
        proofUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [reviewing, cleared, rejected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminWithdrawal:
      type: object
      required: [id, userId, traderId, email, coin, amount, walletAddress, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        traderId:
          type: string
        email:
          type: string
        coin:
          type: string
        amount:
          type: number
        fee:
          type: number
        walletAddress:
          type: string
        status:
          type: string
          enum: [in_transit, cleared, rejected]
        txid:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminUser:
      type: object
      required: [id, email, traderId, balance, kycStatus, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        traderId:
          type: string
        balance:
          type: number
        totalDeposited:
          type: number
        totalInvested:
          type: number
        kycStatus:
          type: string
        role:
          type: string
        guildCode:
          type: ["string", "null"]
        referredBy:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AdminRejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    AdminProcessWithdrawalInput:
      type: object
      required: [txid]
      properties:
        txid:
          type: string
...